fix(options): pass proper loader options to children (#266)
* Pass proper loader `options` to children.

Using just the serialized request means that query objects (that is to say, those which are not strings), get munged if they have anything non-serializable in them. This means passing data to `postcss` via `query` is impossible. To remedy this, we detect when the original request is being processed and apply the original loader transformations instead of whatever came through the serialized version.

* Update `loader-utils` to latest.