docs(README): add example how-to disable `url()` resolving (#677)
* Add array of loaders and options

I was scouring the internet looking for how to get my css url() to resolve when webpack compiles. I couldn't find anything! But, I took a guess at setting `url: false` in the options for my loader and it worked! It would be beneficial to have this documentation on the readme. Thanks for the awesome plugin!

* Update README.md